Completing the chapter, with problems.
Clearly this file as a "collection of tests" will not complete
successfully, since in The Reasoned Schemer, we repeatedly redefine the
`pmembero` function. However, if you pass each function to the REPL,
it works.

I lied. Problem 100 calls a faulty function, `memberrevo`, which
doesn't work. But it is a start.