I could use some feedback on testing poorly formed LCons lists.
The Reasoned Schemer often pushes the boundries of what would be
"normal usage" and ends up creating lists that aren't proper lists.
Normal lists consists of a series of tuples where the "right-side" of
each tuple points to the next element in the series, but the last,
points to null.

In Scheme, you can have the last element's right-side tuple actually
contain a value, this is done with a dot operator:  ( a b . c )

I can't seem to figure out a way to create such behavior for
core.logic, and would love some feedback.